% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_comparison_fxns.R
\name{add_control_values}
\alias{add_control_values}
\title{Add control values}
\usage{
add_control_values(treatment_mean_df)
}
\arguments{
\item{treatment_mean_df}{result of get_treatment_means}
}
\value{
treatment_mean_df with columns for control values at every timestep
}
\description{
Adds columns for control values to a treatment mean df
}
