% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{add_eras}
\alias{add_eras}
\title{Add "eras"}
\usage{
add_eras(dat)
}
\arguments{
\item{dat}{Dataset with column "period"}
}
\value{
dat with column "era" added
}
\description{
Breaking the time series into 4 chunks: prior to 1996, 1996-ca 2010, 2010-2015, post plot switch
}
