% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{add_plot_types}
\alias{add_plot_types}
\title{Add plot types}
\usage{
add_plot_types(dat)
}
\arguments{
\item{dat}{A dataset with column Plot}
}
\value{
dat with plot types added
}
\description{
Add plot types
}
