% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{add_temp_treatments}
\alias{add_temp_treatments}
\title{Add temporary treatments}
\usage{
add_temp_treatments(dat)
}
\arguments{
\item{dat}{dataframe iwth plot_type, period}
}
\value{
dat with column temp_plot_type for what treatment a plot had at each timestep
}
\description{
Add temporary treatments
}
