% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRepeatVisits.R
\name{calcRepeatVisits}
\alias{calcRepeatVisits}
\title{Calculate the number of repeat visits to grid cells across a taxonomic group}
\usage{
calcRepeatVisits(inPath, taxon, res, start, end, output)
}
\arguments{
\item{inPath}{String. Location of the data obtained using extract_records}

\item{taxon}{string. Taxonomic group. Must match a file name (without "_raw_data.rdata") in inPath.}

\item{res}{Numeric. Spatial resolution for the analysis in decimal degrees.}

\item{start}{Numeric. Earliest year cutoff for analysis.}

\item{end}{Numeric. Final year cutoff for analysis}

\item{output}{String. One of: "map" to get a map of the no. records per cell, "hist" to get a histrogram of repeat visits, or other to get the raw values (no. repeat visits per cell).}
}
\description{
This function can be used to assess, for a chosen taxonomic group, the number of records per grid cell.
}
