% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipDat.R
\name{clipDat}
\alias{clipDat}
\title{\code{clipDat} - Extract a geographic subset of your occurrence records.}
\usage{
clipDat(dat, shp)
}
\arguments{
\item{dat}{String. Outputs from \code{formatData}.}

\item{shp}{String. A polygon which will be used to clip the data.}
}
\value{
a data.frame with the new subset of occurrence records.
}
\description{
Extracts all records that fall within a provided polygon (e.g. a country boundary).
}
