% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R, R/createPresAb.R
\name{createPresAb}
\alias{createPresAb}
\title{Create pseudo absences and combine with the scheme occurrence data.}
\usage{
createPresAb(
  inPath,
  taxon,
  species,
  minYear,
  maxYear,
  nAbs,
  matchPres = FALSE,
  recThresh
)

createPresAb(
  inPath,
  taxon,
  species,
  minYear,
  maxYear,
  nAbs,
  matchPres = FALSE,
  recThresh
)
}
\arguments{
\item{inPath}{String. Path to species data extracted using extract_records.}

\item{taxon}{String. Taxonomic group. Must match one of the groups in inPath.}

\item{species}{String. Focal species' name (must be in chosen taxonomic group)}

\item{minYear}{Numeric.}

\item{maxYear}{Numeric. Threshold number of records below which species' are dropped.}

\item{nAbs}{Numeric. Number of pseudo absences to produce. Note that this is overriden if matchPres = TRUE.}

\item{matchPres}{Logical. If true, will produce absences such that there are an equal number to presences. Defaults to false}

\item{recThresh}{Numeric. Lower threshold number of records; species with fewer records are dropped.}

\item{matchpres}{Logical. If TRUE this overrides nAbs and creates pseudo absences in equal number to the occurrence data.}

\item{res}{Numeric. Spatial resolution in decimal degrees (defaults to 0.5)}
}
\value{
A list with n elements, where n is the number of species in the group. Each element contains two
dataframes, one with the coordinates of the presence data, and a second with the coordinates
of the pseudo absences. The coordinate reference system is OSGB36.
}
\description{
This function creates pseudo absences according to the "target group" approach (Phillips 2009).
The aim is to generate pseudo absences with similar environmental bias as the occurrence data.
Pseudo absences are created for individual species, but the function works at the group level
because absences are inferred from the presences of non-focal species in the same group.

Create environmental layers in ascii for a specified region and spatial resolution
}
\examples{

}
