% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodent_energy_mods.R
\name{est_glm_ilink}
\alias{est_glm_ilink}
\title{Get (invlink) estimates from GLM}
\usage{
est_glm_ilink(glm_mod, glm_dat)
}
\arguments{
\item{glm_mod}{model}

\item{glm_dat}{newdat}
}
\value{
estimates
}
\description{
Get (invlink) estimates from GLM
}
