% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_records.R
\name{extract_records}
\alias{extract_records}
\title{extract records from GBIF and format them for use in soaR}
\usage{
extract_records(
  continent = NULL,
  country = NULL,
  year,
  taxon,
  write_output,
  output_path,
  output_name
)
}
\arguments{
\item{continent}{String. E.g. "south_america", "europe". Do not specify unless country is NULL.}

\item{country}{string. E.g. "BR" for Brasil, "AR" for Argentina, and "CL" for Chile}

\item{year}{string. year(s) for which to extract data. For a range, format as e.g. "1995,2010". For a single year, as e.g. "2005".}

\item{taxon}{numeric. GBIF taxonomic key.}

\item{write_output}{logical}

\item{output_path}{string. Directory in which to save outputs.}

\item{output_name}{string. Name of file to be saved if write_output = TRUE.}
}
\description{
This function is used to scrape GBIF for occurrence records for a chosen taxonomic group, on a chosen continent.
}
