% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{formatData}
\alias{formatData}
\title{Format scheme records for use in soaR}
\usage{
formatData(inPath, taxa, write, degrade = TRUE, outPath)
}
\arguments{
\item{inPath}{String. Path to the scheme data directory}

\item{taxa}{string. Taxonomic group. There must be a file in inPath that contains the taxonomic group name.}

\item{write}{Logical. Write the outputs to file?}

\item{degrade}{Logical. Whether or not to degrade the data from all records to uniqe species/ location records.}

\item{outPath}{String. Where to store the outputs.}
}
\value{
A dataframe with four columns: species, year, eastings and northings.
}
\description{
This function subsets scheme data by years, and converts UK or NI grid cells to eastings and northings
on the OSGB36 coordinate reference system.
}
