% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fxns.R
\name{get_compensation}
\alias{get_compensation}
\title{Get compensation}
\usage{
get_compensation(treatl = NULL)
}
\arguments{
\item{treatl}{result of \code{get_treatment_means}}
}
\value{
df with columns \code{smgran_comp}, compensation as defined above, and \code{smgran_comp_ma}, the 6-month moving average.
}
\description{
Calculate degree to which small granivores on exclosure plots compensate for kangaroo rat removal. Compensation calculated as (SmgranEnergy_Exclosures - SmgranEnergy_Controls) / DipoEnergy_Controls.
}
