% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fxns.R
\name{get_dipo_c}
\alias{get_dipo_c}
\title{Get Dipo data for analysis.}
\usage{
get_dipo_c(treatl = NULL)
}
\arguments{
\item{treatl}{result of \code{get_treatment_means}}
}
\value{
\code{treatl} with added columns dipo_prop, the proportion of total energy accounted for by all Dipodomys, and dipo_prop_ma, the 6-month moving average of dipo_prop_ma.


}
\description{
Get kangaroo rat proportional energy use on control plots in each time step for each treatment type.
}
