% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fxns.R
\name{get_e_ratio}
\alias{get_e_ratio}
\title{Get EE energy use as fraction of CC}
\usage{
get_e_ratio(treatl = NULL)
}
\arguments{
\item{treatl}{df}
}
\value{
df
}
\description{
Get EE energy use as fraction of CC
}
