% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_erodium_plot}
\alias{get_erodium_plot}
\title{Get Erodium data (plot level)}
\usage{
get_erodium_plot(census_season = "winter")
}
\arguments{
\item{census_season}{"winter"}
}
\value{
df
}
\description{
Includes rows for empty plots.
}
