% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_erodium_treatment}
\alias{get_erodium_treatment}
\title{Get Erodium data (treatment level)}
\usage{
get_erodium_treatment(census_season = "winter")
}
\arguments{
\item{census_season}{"winter"}
}
\value{
df
}
\description{
To use these data, note that there are 5 EE and 4 CC plots. The prop data is fine but you want to standardize the totals.
}
