% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fxns.R
\name{get_pb}
\alias{get_pb}
\title{Get PB data for analysis}
\usage{
get_pb(treatl = NULL)
}
\arguments{
\item{treatl}{result of \code{get_treatment_means}}
}
\value{
\code{treatl} with added columns pb_prop, the proportion of total energy accounted for by C. baileyi, and pb_prop_ma, the 6-month moving average of pb_prop.
}
\description{
Calculate the proportion of total energy use accounted for by C. baileyi (species code PB) in each time step for each treatment type.
}
