% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_plants}
\alias{get_plants}
\title{Get plot-level plant data}
\usage{
get_plants(census_season = "winter")
}
\arguments{
\item{census_season}{"winter" or "summer"}
}
\value{
wide dataframe of species abundance per plot, including records for empty plots.
}
\description{
For winter and summer censuses for CC, EE plots from 1988-2019, all plots are censused and all plots have 16 quadrats censused. There are however plots that had no individuals. This function adds those in as empties (0 abundance)
}
