% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_plants_annual_ldats}
\alias{get_plants_annual_ldats}
\title{Get Portal plants ready for ldats}
\usage{
get_plants_annual_ldats(census_season = "winter", plot_type = "CC")
}
\arguments{
\item{census_season}{"winter" or summmer}

\item{plot_type}{"CC", "CE", "EE", "EC"}
}
\value{
ready for ldats
}
\description{
Get Portal plants ready for ldats
}
