% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_plants_plots}
\alias{get_plants_plots}
\title{Get plot-level plant data}
\usage{
get_plants_plots(census_season = "winter")
}
\arguments{
\item{census_season}{"winter" or "summer"}
}
\value{
wide dataframe
}
\description{
Get plot-level plant data
}
