% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_plot_abundance_matrix}
\alias{get_plot_abundance_matrix}
\title{Get plot-level plant abundance matrix}
\usage{
get_plot_abundance_matrix(census_season = "winter")
}
\arguments{
\item{census_season}{"winter" or "summer"}
}
\value{
df
}
\description{
Includes rows for empty plots.
}
