% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_plot_totals}
\alias{get_plot_totals}
\title{Get plot totals}
\usage{
get_plot_totals(clean = F, currency = "energy")
}
\arguments{
\item{clean}{passed to portalr, whether to use only qa data or not
#' @param currency "energy" or "biomass"}
}
\value{
data
}
\description{
Quick wrapper for get_rodent_data.
}
