% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_rodent_data}
\alias{get_rodent_data}
\title{Get rodent data}
\usage{
get_rodent_data(return_plot = F, currency = "energy", clean = F)
}
\arguments{
\item{return_plot}{Return plot level energy use or return treatment level. If TRUE, returns plot level totals. If F, returns mean per treatment per period}

\item{currency}{"energy" (default) or "abundance"}

\item{clean}{passed to portalr, whether to use only qa data or not}
}
\value{
data
}
\description{
Core function for getting rodent data.
}
\details{
Gives data from February 1988 until January 2020.
}
