% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_treatment_abundance_matrix}
\alias{get_treatment_abundance_matrix}
\title{Get treatment-level plant abundance matrix}
\usage{
get_treatment_abundance_matrix(census_season = "winter")
}
\arguments{
\item{census_season}{"winter" or "summer"}
}
\value{
df
}
\description{
Includes rows for empty treatments, columns for number of plots, total abundance, and number of empty plots.
}
\details{
Note that there are 4 CC and 5 EE plots. Before using, you want to standardize.
}
