% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimData.R
\name{getClimData}
\alias{getClimData}
\title{Process climate data for use in modelling species' distributions}
\usage{
getClimData(res = 0.5, xMin, xMax, yMin, yMax, inPath, outPath, bioClim, write)
}
\arguments{
\item{res}{Numeric. Spatial resolution in decimal degrees (defaults to 0.5)}

\item{xMin}{Numeric. Min x coordinate.}

\item{xMax}{Numeric. Max x coordinate.}

\item{yMin}{Numeric. Min y coordinate.}

\item{yMax}{Numeric. Max y coordinate.}

\item{inPath}{String. Path to climate data.}

\item{outPath}{Logical. If true the climate data are written as .asc files to outPath.}

\item{bioClim}{Logical. If true, getClimData will automatically name the data layers.}
}
\description{
Create environmental layers in ascii for a specified region and spatial resolution
}
\examples{

}
