% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{getSkill}
\alias{getSkill}
\title{Extract skill of SDMs.}
\usage{
getSkill(inPath, group)
}
\arguments{
\item{inPath}{String. Location of the fitSDM outputs.}

\item{group}{String. Taxonomic group. Must partially match a filename in inPath.}
}
\value{
A dataframe with columns for species and the skill of each type of model for that species.
}
\description{
Extracts the AUC score for each species/ model combo in a group.
}
