% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{getSpNames}
\alias{getSpNames}
\title{Extract the names of all species in a taxonomic group for which there are scheme data.}
\usage{
getSpNames(inPath, taxa)
}
\arguments{
\item{inPath}{String. Path to outputs of formatData.}

\item{taxa}{string. Taxonomic group. There must be a file in inPath that contains the taxonomic group name.}
}
\value{
A vector of species names.
}
\description{
Extract the names of all species in a taxonomic group for which there are scheme data.
}
