% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{list_plot_types}
\alias{list_plot_types}
\title{List which plots are which treatments}
\usage{
list_plot_types(use_pre_switch = F)
}
\arguments{
\item{use_pre_switch}{use treatments to 2015?}
}
\value{
dataframe of plots & treatments
}
\description{
For checking
}
