% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{maopts}
\alias{maopts}
\title{Moving average with default options}
\usage{
maopts(x, n = 6, type = "s")
}
\arguments{
\item{x}{the thing to average}

\item{n}{window length, default 6}

\item{type}{type, default "s"}
}
\value{
movavg
}
\description{
Moving average with default options
}
