% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{modelAverage}
\alias{modelAverage}
\title{Create ensemble predictions.}
\usage{
modelAverage(inPath, outPath, skillDat, species)
}
\arguments{
\item{inPath}{String. Location of the fitSDM outputs.}

\item{outPath}{String. Where to store ensemble models.}

\item{species}{String. Species name.}

\item{skillDat.}{String. getSkill outputs for the chosen group.}
}
\value{
A raster layer with the AUC-weighted average probabilities of occurrence predicted by models.
}
\description{
AUC-weighted average predictions from the models fitted for a species using fitSDM.
}
