\name{NAME}
\alias{NAME}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Dummy function to keep the package checker quiet.
}
\description{
Dummy function, please ignore!
}
\usage{
NAME()
}
\value{
\code{NULL}
}

\examples{
## This is a dummy function included ONLY to get around
## a limitation of the parser used by the package checker.

## The function is currently defined as
function () 
NULL
}
\keyword{programming}

