% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{optOccThresh}
\alias{optOccThresh}
\title{Create binary presence/ absence predictions from probabilities of occurrence.}
\usage{
optOccThresh(inPath, group, species, outPath, map, presAb)
}
\arguments{
\item{inPath}{String. Path to the ensemble model outputs.}

\item{group}{string. Taxonomic group. There must be a file in inPath.}

\item{species}{String. There must be a file in inPath called /<group>/<species>.asc.}

\item{outPath}{String. Where to store the outputs.}

\item{map}{Logical. If true writes an ascii file to outPath.}

\item{presAb}{String. Presence/ pseudo absence data created by createPresAb for the chosen group.}
}
\value{
A dataframe with four columns: Threshold probability of occurrence, species and group.
Optionally a raster can be written to file.
}
\description{
This function optimises a threshold probability above which a grid cell is classed as a presence,
and below which it is classed as an absence. Optimisation is achieved
by maximising the proportion of data correctly classified.
}
