% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_records_spatial.R
\name{plot_records_spatial}
\alias{plot_records_spatial}
\title{Visualize spatial variation in recorder activity}
\usage{
plot_records_spatial(inPath, taxon, res = 0.5, n_periods = 2)
}
\arguments{
\item{inPath}{String. Directory in which the extract_records outputs were saved.}

\item{taxon}{String. Taxonomic group for which to plot records. Must match one of the groups passed to extract_records().}

\item{res}{Numeric. Spatial resolution (decimal degrees) to plot records at. Defaults to 0.5, i.e. half a degree.}

\item{n_periods}{Numeric. How many periods to split the data in to. Defaults to two.}
}
\description{
Map the spatial distribution of records in a chosen number of time-periods
}
\examples{

}
