% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecords.R
\name{plot_records_temporal}
\alias{plot_records_temporal}
\title{Summarise temporal variation in recorder activity}
\usage{
plot_records_temporal(inPath)
}
\arguments{
\item{inPath}{String. Directory in which the extract_records outputs were saved.}
}
\description{
Plot a time series of the number of records (1/year) for taxa selected using extract_records
}
