% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treatments.R
\docType{data}
\name{plot_treatments}
\alias{plot_treatments}
\title{Table of plot treatments over switches}
\format{
A dataframe
}
\usage{
plot_treatments
}
\description{
Reconstructed from Portal documentation (https://github.com/weecology/PortalData) and code supporting Christensen et al. 2019 (https://github.com/emchristensen/PlotSwitch/releases/tag/v2.0.0)
}
\keyword{datasets}
