% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{plots_to_treatment_means}
\alias{plots_to_treatment_means}
\title{Go from plot level totals to treatment means}
\usage{
plots_to_treatment_means(plot_level_totals, currency)
}
\arguments{
\item{plot_level_totals}{plots}

\item{currency}{"energy" or "abundance"}
}
\value{
treatment means
}
\description{
Go from plot level totals to treatment means
}
