% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prioSpecies}
\alias{prioSpecies}
\title{Name of species one or more of England, Scotland, Wales or Northern Ireland's priority lists.}
\format{A data frame with one column: species name.}
\source{
JNCC taxon designations https://hub.jncc.gov.uk/assets/478f7160-967b-4366-acdf-8941fd33850b.
             Also includes Biological Records Centre concept codes for species which have them.
}
\usage{
prioSpecies
}
\description{
Name of species one or more of England, Scotland, Wales or Northern Ireland's priority lists.
}
\examples{
\dontrun{
 prioSpecies
}
}
\keyword{datasets}
