% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordBreakdown.R
\name{recordBreakdown}
\alias{recordBreakdown}
\title{Break down the data by country or basis of record}
\usage{
recordBreakdown(inPath, taxon, start, end, output)
}
\arguments{
\item{inPath}{String. Location of extract_records outputs.}

\item{taxon}{String. Must match one of the taxon names in inPath.}

\item{start}{Numeric. Lower bound for analysis (year).}

\item{end}{Numeric. Upper bound for analysis (year).}

\item{output}{String. One of "country" or "basis".}
}
\description{
This function splits the data obtained using extract_data by country or basis of record (e.g. fossil specimen, preserved specimen)
}
