% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{remove_switch}
\alias{remove_switch}
\title{Remove switch}
\usage{
remove_switch(a_df)
}
\arguments{
\item{a_df}{df with cols plot_type, era}
}
\value{
df filterd to longterm plots and with the final era changed to c_post_cpt
}
\description{
Remove switch
}
