% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTelfer.R
\name{runTelfer}
\alias{runTelfer}
\title{Assess relative species distribution changes in a taxonomic group}
\usage{
runTelfer(inPath, taxon, res = 0.5, p1, p2)
}
\arguments{
\item{inPath}{String. Directory in which the extract_records outputs were saved.}

\item{taxon}{String. Taxonomic group for which to run the Telfer analysis.}

\item{res}{Numeric. Spatial resolution (decimal degrees) to conduct the analysis at. Defaults to 0.5, i.e. half a degree.}

\item{p1}{Numeric or numeric vector. Years in the first period (e.g. 1950:1990).}

\item{p2}{Numeric or numeric vector. Years in the second period.}
}
\description{
Telfer et al. (2002) method for calculating change in distributions between two periods.
}
\examples{

}
