% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingOverlap.R
\name{samplingOverlap}
\alias{samplingOverlap}
\title{Calculate the geographical overlap of sampling in two periods for a given taxa}
\usage{
samplingOverlap(inPath, taxon, res = 0.5, p1, p2)
}
\arguments{
\item{inPath}{String. Directory in which the extract_records outputs were saved.}

\item{taxon}{String. Taxonomic group of interest.}

\item{res}{Numeric. Spatial resolution (decimal degrees) to conduct the analysis at. Defaults to 0.5, i.e. half a degree.}

\item{p1}{Numeric or numeric vector. Years in the first period (e.g. 1950:1990).}

\item{p2}{Numeric or numeric vector. Years in the second period.}
}
\description{
Calculates the number of cells, at a specified resolution, sampled in both of two user-chosen time periods.
}
\examples{

}
