% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{stackPreds}
\alias{stackPreds}
\title{Create taxon-specific maps of species richness.}
\usage{
stackPreds(inPath, group, write = TRUE, outPath)
}
\arguments{
\item{inPath}{String. Location of the modelAverage outputs.}

\item{group}{String. Taxonomic group.}

\item{outPath}{String. Where to store the outputs if write = TRUE.}

\item{write.}{Logical. Should the outputs be written to file?}
}
\value{
A raster layer with the AUC-weighted average probabilities of occurrence predicted by models.
}
\description{
Sum the predicted ensemble probabilities of occurrence for species in the chosen group.
}
