% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{add_eras}
\alias{add_eras}
\title{Add "eras"}
\usage{
add_eras(dat)
}
\arguments{
\item{dat}{Dataset with column "period"}
}
\value{
dat with column "era" added
}
\description{
Breaking the time series into 3 chunks: Jan 1988-June 1997, July 1997-January 2010, January 2010-January 2020
}
