% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMFunctions.R
\name{formatData}
\alias{formatData}
\title{Format scheme records for use in soaR}
\usage{
formatData(occ, taxa, minYear, maxYear, degrade = TRUE)
}
\arguments{
\item{occ}{String. data.frame with occurrence records. Must have columns "YEAR", "CONCEPT" (species name) and one of "TO_GRIDREF" or "SQ_1KM".}

\item{taxa}{string. Taxonomic group. There must be a file in inPath that contains the taxonomic group name.}

\item{minYear}{Numeric. Drop all records prior to this year.}

\item{maxYear}{Numeric. Drop all records later than this year.}

\item{degrade}{Logical. Whether or not to degrade the data from all records to uniqe species/ location records.}
}
\value{
A dataframe with four columns: species, year, eastings and northings.
}
\description{
This function subsets scheme data by years, and converts UK or NI grid cells to eastings and northings
on the OSGB36 coordinate reference system.
}
