% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fxns.R
\name{get_e_ratio}
\alias{get_e_ratio}
\title{Get EE energy use as fraction of CC}
\usage{
get_e_ratio(treatl = NULL)
}
\arguments{
\item{treatl}{result of \code{get_treatment_means}}
}
\value{
dataframe with column \code{total_e_rat}, the ratio of \code{treatl$total_e} on exclosures relative to controls in each timestep, and \code{total_e_rat_ma}, the moving average.
}
\description{
For each timestep, calculate ratio of total granivore energy use on exclosure plots to control plots.
}
