% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_plot_totals}
\alias{get_plot_totals}
\title{Get plot totals}
\usage{
get_plot_totals(use_pre_switch = F)
}
\arguments{
\item{use_pre_switch}{use pre switch T/F}
}
\value{
data
}
\description{
Quick wrapper for get_rodent_data.
}
