% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_rodent_data}
\alias{get_rodent_data}
\title{Get rodent data}
\usage{
get_rodent_data(use_christensen_plots = F, return_plot = F, use_pre_switch = F)
}
\arguments{
\item{use_christensen_plots}{Early in development I was working from the plots used in Christensen (2019 ProcB). Defaults F}

\item{return_plot}{Return plot level energy use or return treatment level. If TRUE, returns plot level totals. If F, returns mean per treatment per period.}

\item{use_pre_switch}{Use data up to the treatment switch in 2015? If yes, allows for more plots of each treatment type.}
}
\value{
data
}
\description{
Core function for getting rodent data.
}
\details{
Gives data from February 1988 until either March 2015 or May 2019, depending on use_pre_switch.
}
