% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_treatment_means}
\alias{get_treatment_means}
\title{Get treatment means}
\usage{
get_treatment_means(clean = F)
}
\arguments{
\item{clean}{passed to portalr, whether to use only qa data or not}

\item{currency}{"energy" or "biomass"}
}
\value{
data
}
\description{
Quick wrapper for get_rodent_data.
}
