% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_treatment_means}
\alias{get_treatment_means}
\title{Get treatment means}
\usage{
get_treatment_means(use_pre_switch = F)
}
\arguments{
\item{use_pre_switch}{use pre switch T/F}
}
\value{
data
}
\description{
Quick wrapper for get_rodent_data.
}
